SingleBusiness Component: created utility functions folder and moved convert time function still commented out in SingleBusiness.js
