+++ client/store/singleBusiness.js
@@ -1,17 +1,9 @@
 import axios from "axios";
 
 //Action type
-const GET_BUSINESS = "GET_BUSINESS";
 const GET_BUSINESS_DETAILS = "GET_BUSINESS_DETAILS";
 
 //Action Creator
-export const getBusiness = (business) => {
-  return {
-    type: GET_BUSINESS,
-    business,
-  };
-};
-
 export const getBusinessDetails = (business) => {
   return {
     type: GET_BUSINESS_DETAILS,
@@ -20,17 +12,6 @@ export const getBusinessDetails = (business) => {
 };
 
 //thunk
-export const fetchSingleBusiness = (id) => {
-  return async (dispatch) => {
-    try {
-      const { data } = await axios.get(`/api/maps/${id}`);
-      dispatch(getBusiness(data));
-    } catch (error) {
-      console.log(error);
-    }
-  };
-};
-
 export const fetchSingleBusinessDetails = (id) => {
   return async (dispatch) => {
     try {
@@ -45,8 +26,6 @@ export const fetchSingleBusinessDetails = (id) => {
 const initialState = {}; //object for single business
 export default function businessReducer(state = initialState, action) {
   switch (action.type) {
-    case GET_BUSINESS:
-      return action.business;
     case GET_BUSINESS_DETAILS:
       return action.business;
     default:
