+++ client/utilityFunctions.js
@@ -1,0 +1,28 @@
+//File for all utility functions used elsewhere in the app
+
+//function used to convert military time to standard time
+export function convertTime(time) {
+  let numValue = Number(time) / 100;
+  let endDigits = '00'
+  let AmOrPm; 
+  if (!Number.isInteger(numValue)){
+      let timeArr = String(numValue).split('.') 
+      numValue = Number(timeArr[0])
+      endDigits = timeArr[1].concat('0')
+  }
+  
+  if (numValue === 24){
+    numValue = 12
+    AmOrPm = 'am'
+  }else if (numValue >= 12){
+    AmOrPm = 'pm'
+  }else{
+    AmOrPm = 'am'
+  }
+
+  if (endDigits.length > 2){
+    endDigits = endDigits.slice(0,2)
+  }
+  let finalTime = AmOrPm === "pm" && numValue !== 12 ? `${String(numValue - 12)}:${endDigits} ${AmOrPm}`: `${String(numValue)}:${endDigits} ${AmOrPm}`;
+  return finalTime;
+}
